+++ packages/vsce/__tests__/__e2e__/07_ResourceInspectorForPrograms.test.ts
@@ -80,7 +80,7 @@ describe("Show Resource Inspector For Programs", () => {
 
   it("Open Resource Inspector on C128N", async () => {
     await C128NProgram?.click();
-    await runCommandFromCommandPalette(">IBM CICS for Zowe Explorer: Inspect Resource");
+    await runCommandFromCommandPalette(">Zowe Explorer for IBM CICS Transaction Server: Inspect Resource");
 
     resourceInspectorPanel = new BottomBarPanel();
     await resourceInspectorPanel.openTab("CICS Resource Inspector");