+++ packages/vsce/src/constants/CICS.messages.ts
@@ -11,7 +11,7 @@
 
 import { IMessageDefinition } from "@zowe/imperative";
 
-export const CICSMessages: { [key: string]: IMessageDefinition; } = {
+export const CICSMessages: { [key: string]: IMessageDefinition } = {
   zoweExplorerNotFound: {
     message: "Zowe Explorer was not found: Please ensure Zowe Explorer v2.0.0 or higher is installed",
   },
@@ -21,11 +21,12 @@ export const CICSMessages: { [key: string]: IMessageDefinition; } = {
   },
 
   notInitializedCorrectly: {
-    message: "IBM CICS for Zowe Explorer was not initialized correctly.",
+    message: "Zowe Explorer for IBM CICS Transaction Server was not initialized correctly.",
   },
 
   incorrectZoweExplorerVersion: {
-    message: `Zowe Explorer was not found: either it is not installed or you are using an older version without extensibility API. ` +
+    message:
+      `Zowe Explorer was not found: either it is not installed or you are using an older version without extensibility API. ` +
       `Please ensure Zowe Explorer v2.0.0-next.202202221200 or higher is installed`,
   },